requirements.txt
requests~=2.32.3
pillow~=10.3.0
aiohttp~=3.9.5
beautifulsoup4~=4.12.3
selenium~=4.22.0
python-dotenv~=1.0.1